+++ history.md
@@ -3,6 +3,7 @@
 
 ## v0.2.1 2022-01-08 ##
 * Fixed issues of utilities for preparing input files
+* Added option for skipping to integrate mtx into finalized prediction 
 
 ## v0.2.0 2021-12-14 ##
 * Added multi-tasking module for processing taxa in parallel